Reverted part of the MenuBuilder commit
Interfaces should only be used on methods NEEDED by ALL instances of
the interface. getReturnImage() should not be called by an outside
class as it is not NEEDED on ALL instances of the interface, so thereby
should not be part of the interface. In fact, none of the code
functionality changed with its removal.