Another revamp of the MSToolkit, now FULLY USABLE!
For any questions/bugs/comments, please ask me. Now I can start working
on the REAL upgrades system, not just the toolkit. Also, all of the
classes in the impl subpackage are implementation dependent, which
means all of the Files in the impl subpackage will change if we move to
a different implementation, and none of the files in the other packages
will change, as they do not rely on the implementation. However, one
exception to this is the MSImage class, which uses the awt class
BufferedImage.